REFACTOR: Extract the code of anonymous classes using `nikic/php-parser`
This adds an additional dependency into the project but serves the
purpose of fully qualifying names in the extracted code. This way the
resolved anonymous classes can be used in different namespaces.

Currently `function` and `const` names are not resolved as there is
always the naming-conflict to the root namespace. This might be solvable
by checking against the reflection of the transformed class.